+++ src/pages/dashboard/components/DashboardFilter.js
@@ -171,7 +171,7 @@ const DashboardFilter = ({ filterQry, setChartData, setTotalLoans, filterType, f
     key: 'range'
   });
 
-  const { data: zones = []} = useQuery('zones', () => {return getZones()}, {refetchOnWindowFocus: false})
+  const { data: zones = []} = useQuery('zones', () => {return getZones(1)}, {refetchOnWindowFocus: false})
 
   const onDatePickerChange = ({ range }) => {
     setDateRange(range)
@@ -286,14 +286,14 @@ const DashboardFilter = ({ filterQry, setChartData, setTotalLoans, filterType, f
     getLoanStats(qry)
       .then(data => {
         let cdata = [
-          { name: 'Submitted', count: data?.data?.submitted_count },
-          { name: 'Pending Review', count: data?.data?.loan_review_count },
-          { name: 'Pending Approval', count: data?.data?.loan_approval_count || 0, amount: data?.data?.amount_requested },
-          { name: 'Approved', count: data?.data?.approved_count, amount: data?.data?.amount_approved },
-          { name: 'Disb. Approval', count: data?.data?.disbursement_approval_count || 0, amount: data?.data?.amount_disbursement_approval },
-          { name: 'Disb. Approved', count: data?.data?.disbursement_approved_count || 0, amount: data?.data?.amount_disbursement_approved },
-          { name: 'Disbursed', count: data?.data?.disbursed_count, amount: data?.data?.actual_amount_disbursed },
-          { name: 'Rejected', count: data?.data?.rejected_count },
+          { name: 'Submitted', count: data?.submitted_count },
+          { name: 'Pending Review', count: data?.loan_review_count },
+          { name: 'Pending Approval', count: data?.loan_approval_count || 0, amount: data?.amount_requested },
+          { name: 'Approved', count: data?.approved_count, amount: data?.amount_approved },
+          { name: 'Disb. Approval', count: data?.disbursement_approval_count || 0, amount: data?.amount_disbursement_approval },
+          { name: 'Disb. Approved', count: data?.disbursement_approved_count || 0, amount: data?.amount_disbursement_approved },
+          { name: 'Disbursed', count: data?.disbursed_count, amount: data?.actual_amount_disbursed },
+          { name: 'Rejected', count: data?.rejected_count },
         ];
         setChartData(cdata);
         let s = 0;