+++ src/pages/dashboard/dashboard.js
@@ -1,4 +1,4 @@
-import { Typography } from '@material-ui/core';
+import { Table, TableBody, TableCell, TableRow, Typography } from '@material-ui/core';
 import Box from '@material-ui/core/Box';
 import Grid from '@material-ui/core/Grid';
 import Paper from '@material-ui/core/Paper';
@@ -11,16 +11,15 @@ import styled from 'styled-components';
 import DashboardFilter from './components/DashboardFilter';
 import LoansTable from './components/LoansTable';
 import LoanStats from './components/LoanStats';
-import { PieChartData, BarChartData, GroupChartData } from './components/MetricsComponents';
+import { PieChartData, BarChartData, GroupChartData, LineChart } from './components/MetricsComponents';
 import Currency from '../../../src/components/Number/Currency';
 import DashCard from '../../components/CommonComponents/Cards/DashCard';
 import LoanBookTable from '../../components/Tables/LoanBookTable';
+import { permissionCheck } from '../../components/UserCan/UserCan'
+import { rulesList } from '../../config/userRules'
 import usePageTitle from '../../hooks/usePageTitle';
-// import { InfoBoxContainer, InfoBoxWrapper } from '../../components/CommonComponents/InfoBox';
-// import { Tooltip, LabelList,Legend, BarChart, CartesianGrid, XAxis, YAxis, Bar, Text } from 'recharts';
 import { getDealerDetails } from '../../services/dealers.service';
-import { getAll_ls1_Metrices, getAll_ls2_Metrices, getAllOmcDpd, getAllRegionDpd } from '../../services/loans.service';
-// import { yellow } from '@material-ui/core/colors';
+import { getAll_ls1_Metrices, getAll_ls2_Metrices, getAllOmcDpd, getAllRegionDpd, getProjectionReport } from '../../services/loans.service';
 
 const currencyFormat = (value) => {
   const money = new Intl.NumberFormat('en-IN', { style: 'currency', currency: 'INR', maximumSignificantDigits: 8 }).format(value)
@@ -38,18 +37,14 @@ function createCustomHTMLContent({ label, data }) {
   `;
 }
 
-// function createCustomHTMLContentforPie(data) {
-//   return `
-//     <div style='padding: 5px; width: 220px'>
-//       <p style='font-size: 12px;'><strong>${data.cust_region}</strong></p>
-//       <table>
-//          <tr><td>Amount</td><td>: <strong> ${currencyFormat(data.od_amount)}</strong></td></tr>
-//       </table>
-//     </div>
-//   `;
-// }
-
-
+const CustomToolTip = (date, day, amount, shortAmt) => {
+  return`
+  <div style='padding: 7px; width: 220px'>
+    <h3>${date} - ${day}</h3>
+    <p style='font-size: 12px; margin-top: 3px'>Due Amount: <strong>${currencyFormat(amount)} (${shortAmt} Cr)</strong></p>
+  </div>
+  `
+}
 
 const arrangeData = (res) => {
   const result = res.reduce((temp, item, i) => {
@@ -73,13 +68,28 @@ const useStyles = makeStyles(theme => ({
       }
     }
   },
-
   dataChart: {
     padding: 10,
     borderRadius: 5,
   },
   noData: {
     padding: 10
+  },
+  tableHead: {
+    '&.MuiTableCell-root.MuiTableCell-body': {
+      backgroundColor: '#eee',
+      position: 'sticky',
+      left: 0
+    },
+    fontSize: 10
+  },
+  item: {
+    fontSize: 10
+  },
+  horizondalTable: {
+    width: '100%',
+    overflow: 'auto',
+    marginTop: 15
   }
 }))
 const DataCharts = styled.div`
@@ -93,7 +103,6 @@ const DataCharts = styled.div`
 const Dashboard = ({ currentUser, dashboardView }) => {
   usePageTitle('Dashboard');
   const classes = useStyles();
-
   const [chartData, setChartData] = useState([{}, {}, {}, {}, {}, {}]);
   const [totalLoans, setTotalLoans] = useState()
   const [ls1_metrices, setLs1Metrices] = useState({});
@@ -107,6 +116,8 @@ const Dashboard = ({ currentUser, dashboardView }) => {
   const [omcData, setOmcData] = useState([]);
   const [RegionData, setRegionData] = useState([]);
   const [filterQry, setFilterQry] = useState();
+  const [LineChartData, setLineChartData] = useState();
+  const [projectionTableData, setProjectionTableData] = useState();
 
   const handleClick = (name) => {
     setSelectedStatsCard(name)
@@ -114,6 +125,20 @@ const Dashboard = ({ currentUser, dashboardView }) => {
   }
 
   useMount(() => {
+    getProjectionReport()
+      .then(data => {
+        setProjectionTableData(data)
+        let testData = data.reduce((temp, item, i) => {
+          if (i === 0) {
+            temp[i] = ['Date', 'Due Amount', { role: 'tooltip', type: 'string', p: { html: true }}];
+          }
+          temp[i+1] = [`${item.due_date.split('-')[0]}/${item.due_date.split('-')[1]}`,item.due_amount ,CustomToolTip(item.due_date, item.short_day, item.due_amount, item.short_amount)]
+          return temp
+        }, [])
+        setLineChartData(testData);
+      })
+      .catch(e => console.log(e))
+
     getAllOmcDpd()
       .then((res) => {
         const result = arrangeData(res)
@@ -125,15 +150,6 @@ const Dashboard = ({ currentUser, dashboardView }) => {
 
     getAllRegionDpd()
       .then((res) => {
-        // let dataRow = []
-        // let dataColumn = []
-        /**
-         * [
-         *  ['', 'x-label 1', 'x-label 2'],
-         *  ['y-label-1', 'x-label-1-value', 'x-label-2-value'],
-         *  ['y-label-2', 'x-label-1-value', 'x-label-2-value']
-         * ]
-         */
         const result = arrangeData(res)
         setRegionData(result)
       })
@@ -141,65 +157,41 @@ const Dashboard = ({ currentUser, dashboardView }) => {
         console.log(e);
       })
 
-    // getLoanStats()
-    //   .then(data => {
-    //     // const data = _countBy(res, item => {
-    //     //   return item.status?.toLowerCase()
-    //     // });
-    //     let cdata = [
-    //       { name: 'Submitted', count: data.submitted_count },
-    //       { name: 'Pending Approval', count: data.loan_approval_count || 0 },
-    //       { name: 'Approved', count: data.approved_count },
-    //       { name: 'Pending Disbursement Approval', count: data.disbursement_approval_count || 0 },
-    //       { name: 'Disbursement Approved',count:data.disbursement_approved_count || 0 },
-    //       { name: 'Disbursed', count: data.disbursed_count },
-    //       { name: 'Rejected', count: data.rejected_count },
-    //     ];
-    //     setChartData(cdata);
-    //   })
-    //   .catch(err => {
-    //     console.log(err);
-    //   })
+    getAll_ls1_Metrices()
+      .then(res => {
+        const result = res[0] || {};
+        setLs1Metrices(result);
+        let overallData = [
+          ['Days', 'Amount', { role: 'tooltip', type: 'string', p: { html: true } }, { role: 'style' }, { role: 'annotation' }],
+          ['<=3 Days', result.lt3_days, currencyFormat(result.lt3_days), '#81B214', currencyFormat(result.lt3_days)],
+          ['4-15 Days', result.gt4lt15_days, currencyFormat(result.gt4lt15_days), '#5C7AEA', currencyFormat(result.gt4lt15_days)],
+          ['15-30 Days', result.gt15lt30_days, currencyFormat(result.gt15lt30_days), '#8236CB', currencyFormat(result.gt15lt30_days)],
+          ['30-60 Days', result.gt30lt60_days, currencyFormat(result.gt30lt60_days), '#FF9300', currencyFormat(result.gt30lt60_days)],
+          ['60-90 Days', result.gt60lt90_days, currencyFormat(result.gt60lt90_days), '#FF6767', currencyFormat(result.gt60lt90_days)],
+          ['>=90 Days', result.gt90_days, currencyFormat(result.gt90_days), '#E02401', currencyFormat(result.gt90_days)],
+        ]
+        setdaysChartData(overallData);
+      })
+      .catch(err => {
+        console.log(err)
+      })
 
-
-    setTimeout(() => {
-      getAll_ls1_Metrices()
-        .then(res => {
-          const result = res[0] || {};
-          setLs1Metrices(result);
-          let overallData = [
-            ['Days', 'Amount', { role: 'tooltip', type: 'string', p: { html: true } }, { role: 'style' }, { role: 'annotation' }],
-            ['<=3 Days', result.lt3_days, currencyFormat(result.lt3_days), '#81B214', currencyFormat(result.lt3_days)],
-            ['4-15 Days', result.gt4lt15_days, currencyFormat(result.gt4lt15_days), '#5C7AEA', currencyFormat(result.gt4lt15_days)],
-            ['15-30 Days', result.gt15lt30_days, currencyFormat(result.gt15lt30_days), '#8236CB', currencyFormat(result.gt15lt30_days)],
-            ['30-60 Days', result.gt30lt60_days, currencyFormat(result.gt30lt60_days), '#FF9300', currencyFormat(result.gt30lt60_days)],
-            ['60-90 Days', result.gt60lt90_days, currencyFormat(result.gt60lt90_days), '#FF6767', currencyFormat(result.gt60lt90_days)],
-            ['>=90 Days', result.gt90_days, currencyFormat(result.gt90_days), '#E02401', currencyFormat(result.gt90_days)],
-          ]
-          setdaysChartData(overallData);
-        })
-        .catch(err => {
-          console.log(err)
-        })
-
-      getAll_ls2_Metrices()
-        .then(res => {
-          const result = res;
-          let total = 0;
-          const colors = ['#4cba6b', '#5899DA', '#E8743B', '#19A979', '#ED4A7B', '#945ECF', '#13A4B4', '#525DF4', '#BF399E', '#6C8893', '#EE6868', '#2F6497', '#f5b04d', '#8a3800', '#008B73', '#42C1AA', '#00A8D2', '#6929c4', '#4589ff']
-          const dataSource = result.map((item, index) => {
-            total += item.od_amount;
-            // createCustomHTMLContentforPie(item.od_amount)
-            return [item.cust_region, item.od_amount, colors[index], currencyFormat(item.od_amount)]
-          });
-          dataSource.length && dataSource.unshift(['Region', 'Amount', { role: 'style' }, { role: 'annotation' },]);
-          setTotalForRegion(currencyFormat(total));
-          setLs2Metrices(dataSource);
-        })
-        .catch(err => {
-          console.log(err)
-        })
-    }, 4000)
+    getAll_ls2_Metrices()
+      .then(res => {
+        const result = res;
+        let total = 0;
+        const colors = ['#4cba6b', '#5899DA', '#E8743B', '#19A979', '#ED4A7B', '#945ECF', '#13A4B4', '#525DF4', '#BF399E', '#6C8893', '#EE6868', '#2F6497', '#f5b04d', '#8a3800', '#008B73', '#42C1AA', '#00A8D2', '#6929c4', '#4589ff']
+        const dataSource = result.map((item, index) => {
+          total += item.od_amount;
+          return [item.cust_region, item.od_amount, colors[index], currencyFormat(item.od_amount)]
+        });
+        dataSource.length && dataSource.unshift(['Region', 'Amount', { role: 'style' }, { role: 'annotation' },]);
+        setTotalForRegion(currencyFormat(total));
+        setLs2Metrices(dataSource);
+      })
+      .catch(err => {
+        console.log(err)
+      })
   });
   useMount(() => {
     getDealerDetails()
@@ -215,18 +207,11 @@ const Dashboard = ({ currentUser, dashboardView }) => {
         ]
         setDealerChartData(dData)
       })
-
       .catch((e) => {
         console.log(e);
       });
   });
 
-
-  // const CustomizedAxisTick = ({ x, y, payload }) => {
-  //   return (
-  //     <Text x={x} y={y} fill='#666' width={70} fontSize='12' fontWeight='bold' textAnchor="middle" verticalAnchor="start">{payload.value}</Text>
-  //   )
-  // }
   return (
     <div style={{ flexGrow: 1 }}>
       {
@@ -251,9 +236,9 @@ const Dashboard = ({ currentUser, dashboardView }) => {
         ) : (
           <>
             <Grid container spacing={2}>
-              <Grid item xs={12}>
-                {
-                  dashboardView === 'LOS' && (
+              {
+                dashboardView === 'LOS' && (
+                  <Grid item xs={12}>
                     <DashboardFilter 
                       filterQry={setFilterQry}
                       setChartData={setChartData}
@@ -261,9 +246,9 @@ const Dashboard = ({ currentUser, dashboardView }) => {
                       filterType='Dashboard'
                       filters={['zone', 'region', 'product', 'period']}
                     />
-                  )
-                }
-              </Grid>
+                  </Grid>
+                )
+              }
               <Grid item xs={12}>
                 {
                   dashboardView === 'LOS' && (
@@ -303,7 +288,7 @@ const Dashboard = ({ currentUser, dashboardView }) => {
                         <BarChartData daysChartData={daysChartData} />
                       </DataCharts>
                     </Grid>
-                    <Grid item md={12} style={{ margin: '10px' }}>
+                    <Grid item md={12} style={{ margin: '10px', marginTop: 17 }}>
                       <DataCharts>
                         {
                           omcData.length ? (
@@ -330,6 +315,52 @@ const Dashboard = ({ currentUser, dashboardView }) => {
                       }
                     </DataCharts>
                   </Grid>
+                  {
+                    permissionCheck(currentUser.role_name, rulesList.projection_report) &&
+                      <>
+                        <Grid item md={12}>
+                          <Paper>
+                            <LineChart chartData={LineChartData} title="Projection Graph" xAxis="Date" />
+                            <div className={classes.horizondalTable}>
+                              <Table>
+                                <TableBody>
+                                  <TableRow>
+                                    <TableCell className={classes.tableHead}><strong>Date</strong></TableCell>
+                                    {
+                                    projectionTableData?.map((item, i) => {
+                                      return(
+                                        <TableCell key={i} className={classes.item}>{item.due_date}</TableCell>
+                                      )
+                                    })
+                                    }
+                                  </TableRow>
+                                  <TableRow>
+                                    <TableCell className={classes.tableHead}><strong>Day</strong></TableCell>
+                                    {
+                                    projectionTableData?.map((item, i) => {
+                                      return(
+                                        <TableCell key={i} className={classes.item}>{item.day}</TableCell>
+                                      )
+                                    })
+                                    }
+                                  </TableRow>
+                                  <TableRow>
+                                    <TableCell className={classes.tableHead}><strong>Due Amount</strong></TableCell>
+                                    {
+                                    projectionTableData?.map((item, i) => {
+                                      return(
+                                        <TableCell key={i} className={classes.item}>{currencyFormat(item.due_amount)}</TableCell>
+                                      )
+                                    })
+                                    }
+                                  </TableRow>
+                                </TableBody>
+                              </Table>
+                            </div>
+                          </Paper>
+                        </Grid>
+                      </>
+                  }
                   <Grid item xs={12}>
                     <LoanBookTable title={'Loan Book'} currentUser={currentUser} />
                   </Grid>