+++ src/pages/transports/components/AddNewVehicleForm.js
@@ -4,20 +4,20 @@ import Divider from '@material-ui/core/Divider';
 import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
 import CloseIcon from '@material-ui/icons/Close';
-import EditIcon from '@material-ui/icons/Edit';
 import NavigateBeforeRoundedIcon from '@material-ui/icons/NavigateBeforeRounded';
-import NavigateNextRounded from '@material-ui/icons/NavigateNextRounded';
 import { makeStyles } from '@material-ui/styles';
 import clsx from 'clsx';
 import { useFormik } from 'formik';
 import { useSnackbar } from 'notistack';
 import React, { useState } from 'react';
+import { useQueryClient } from 'react-query';
 import * as Yup from 'yup';
 import Button from '../../../components/CommonComponents/Button/Button';
+import { ViewData } from '../../../components/CommonComponents/FilePreview';
 import TextInput from '../../../components/TextInput/TextInput';
 import {
   addNewVehicle,
-  updateVehicle,
+  getVehicleInfoFromID,
 } from '../../../services/transports.service';
 
 const useStyles = makeStyles((theme) => ({
@@ -59,6 +59,12 @@ const useStyles = makeStyles((theme) => ({
       backgroundColor: theme.palette.success.dark,
     },
   },
+  items: {
+    '&.MuiGrid-item': {
+      paddingTop: 0,
+      paddingBottom: 2
+    }
+  }
 }));
 
 const AddNewVehicleForm = ({
@@ -70,14 +76,13 @@ const AddNewVehicleForm = ({
   isAdd,
   callback,
 }) => {
+  const queryClient = useQueryClient()
   const [readOnly, setReadOnly] = useState(false);
   const { enqueueSnackbar } = useSnackbar();
   const [loading, setLoading] = useState(false);
+  const [vehicleDetails, setVehicleDetails] = useState()
   const classes = useStyles();
 
-  const handleEdit = () => {
-    setReadOnly(!readOnly);
-  };
   const {
     values,
     errors,
@@ -96,56 +101,33 @@ const AddNewVehicleForm = ({
     }),
     onSubmit: (formData) => {
       setLoading(true);
-      if (isAdd === 'Edit') {
-        updateVehicle(formData, id, trans_id)
-          .then((message) => {
-            setLoading(false);
-            enqueueSnackbar(message, {
-              anchorOrigin: {
-                vertical: 'top',
-                horizontal: 'right',
-              },
-              variant: 'success',
-            });
-
-            setTimeout(() => {
-              window.location.reload();
-            }, 2000);
-          })
-          .catch((e) => {
-            setLoading(false);
-            enqueueSnackbar(e, {
-              anchorOrigin: {
-                vertical: 'top',
-                horizontal: 'right',
-              },
-              variant: 'error',
-            });
+      addNewVehicle(formData, id)
+        .then((message) => {
+          setLoading(false)
+          enqueueSnackbar(message, {
+            anchorOrigin: {
+              vertical: 'top',
+              horizontal: 'right',
+            },
+            variant: 'success',
           });
-      } else {
-        addNewVehicle(formData, id)
-          .then((message) => {
-            enqueueSnackbar(message, {
-              anchorOrigin: {
-                vertical: 'top',
-                horizontal: 'right',
-              },
-              variant: 'success',
-            });
-            setTimeout(() => {
-              window.location.reload();
-            }, 2000);
-          })
-          .catch((e) => {
-            enqueueSnackbar(e, {
-              anchorOrigin: {
-                vertical: 'top',
-                horizontal: 'right',
-              },
-              variant: 'error',
-            });
+          getVehicleInfoFromID(id)
+            .then((data) => {
+              setVehicleDetails(JSON.parse(data?.find(item => item.tt_no === formData.tt_no)?.vehicle_details))
+            })
+            .catch(e => console.log(e))
+        })
+        .catch((e) => {
+          setLoading(false)
+          enqueueSnackbar(e, {
+            anchorOrigin: {
+              vertical: 'top',
+              horizontal: 'right',
+            },
+            variant: 'error',
           });
-      }
+        });
+      // }
     },
   });
   const inputProps = {
@@ -185,64 +167,102 @@ const AddNewVehicleForm = ({
                     />
                   )}
                 </Grid>
+                <Grid item md={6}>
+                  {
+                    loading ? <div style={{height: '6vh', marginTop: 19}}><CircularProgress size={30}/></div> :
+                    <Button style={{marginTop: 21}} variant='contained' type='submit' className={clsx(classes.btn, classes.editButton)} onClick={handleSubmit}>Save</Button>
+                  }
+                </Grid>
               </Grid>
+              <Divider style={{marginTop: 8}} />
             </form>
+            {
+              vehicleDetails &&
+                <Grid container spacing={2} style={{marginTop: 15}}>
+                  <Grid item md={6} className={classes.items}>
+                    <ViewData title="Vehicle Description" value={vehicleDetails?.vehicleClassDescription} style={{marginBottom: 0}} />
+                  </Grid>
+                  <Grid item md={6} className={classes.items}>
+                    <ViewData title="Registration Date" value={vehicleDetails?.registrationDate} style={{marginBottom: 0}} />
+                  </Grid>
+                  <Grid item md={6} className={classes.items}>
+                    <ViewData title="Vehicle Category" value={vehicleDetails?.vehicleCatgory} style={{marginBottom: 0}} />
+                  </Grid>
+                  <Grid item md={6} className={classes.items}>
+                    <ViewData title="Registration No" value={vehicleDetails?.registrationNumber} style={{marginBottom: 0}} />
+                  </Grid>
+                  <Grid item md={6} className={classes.items}>
+                    <ViewData title="Engine No" value={vehicleDetails?.engineNumber} style={{marginBottom: 0}} />
+                  </Grid>
+                  <Grid item md={6} className={classes.items}>
+                    <ViewData title="Chassis No" value={vehicleDetails?.chassisNumber} style={{marginBottom: 0}} />
+                  </Grid>
+                  <Grid item md={6} className={classes.items}>
+                    <ViewData title="No.of Cylinders" value={vehicleDetails?.numberOfCylinders} style={{marginBottom: 0}} />
+                  </Grid>
+                  <Grid item md={6} className={classes.items}>
+                    <ViewData title="Maker Description" value={vehicleDetails?.makerDescription} style={{marginBottom: 0}} />
+                  </Grid>
+                  <Grid item md={6} className={classes.items}>
+                    <ViewData title="Fuel Type" value={vehicleDetails?.fuelDescription} style={{marginBottom: 0}} />
+                  </Grid>
+                  <Grid item md={6} className={classes.items}>
+                    <ViewData title="Model" value={vehicleDetails?.makerModel} style={{marginBottom: 0}} />
+                  </Grid>
+                  <Grid item md={6} className={classes.items}>
+                    <ViewData title="Capacity" value={vehicleDetails?.cubicCapacity} style={{marginBottom: 0}} />
+                  </Grid>
+                  <Grid item md={6} className={classes.items}>
+                    <ViewData title="Color" value={vehicleDetails?.color} style={{marginBottom: 0}} />
+                  </Grid>
+                  <Grid item md={6} className={classes.items}>
+                    <ViewData title="Owner Name" value={vehicleDetails?.ownerName} style={{marginBottom: 0}} />
+                  </Grid>
+                  <Grid item md={6} className={classes.items}>
+                    <ViewData title="Insurance Upto" value={vehicleDetails?.insuranceUpto} style={{marginBottom: 0}} />
+                  </Grid>
+                  <Grid item md={6} className={classes.items}>
+                    <ViewData title="Insurance Policy No" value={vehicleDetails?.insurancePolicyNumber} style={{marginBottom: 0}} />
+                  </Grid>
+                  <Grid item md={6} className={classes.items}>
+                    <ViewData title="Fitness Upto" value={vehicleDetails?.fitnessUpto} style={{marginBottom: 0}} />
+                  </Grid>
+                  <Grid item md={6} className={classes.items}>
+                    <ViewData title="Manufactured Month Year" value={vehicleDetails?.manufacturedMonthYear} style={{marginBottom: 0}} />
+                  </Grid>
+                  <Grid item md={6} className={classes.items}>
+                    <ViewData title="Insurance Company" value={vehicleDetails?.insuranceCompany} style={{marginBottom: 0}} />
+                  </Grid>
+                  <Grid item md={6} className={classes.items}>
+                    <ViewData title="PUC No" value={vehicleDetails?.pucNumber} style={{marginBottom: 0}} />
+                  </Grid>
+                  <Grid item md={6} className={classes.items}>
+                    <ViewData title="PUC Exp Date" value={vehicleDetails?.pucExpiryDate} style={{marginBottom: 0}} />
+                  </Grid>
+                  <Grid item md={6} className={classes.items}>
+                    <ViewData title="Blacklist Status" value={vehicleDetails?.blackListStatus} style={{marginBottom: 0}} />
+                  </Grid>
+                  <Grid item md={6} className={classes.items}>
+                    <ViewData title="Rc Status" value={vehicleDetails?.rcStatus} style={{marginBottom: 0}} />
+                  </Grid>
+                  <Grid item md={6} className={classes.items}>
+                    <ViewData title="Rc Mob No" value={vehicleDetails?.rcMobileNo} style={{marginBottom: 0}} />
+                  </Grid>
+                  <Grid item md={6} className={classes.items}>
+                    <ViewData title="Blacklist Info" value={vehicleDetails?.blackListInfo} style={{marginBottom: 0}} />
+                  </Grid>
+                </Grid>
+            }
           </Box>
-        </div>
-      </div>
-      <div className={classes.actionFooter}>
-        <Divider />
-        <div className={classes.actionButtonsWrapper}>
           <Button
             variant='outlined'
             startIcon={<NavigateBeforeRoundedIcon />}
             disabled={loading}
-            onClick={callback}
+            onClick={() => {callback(); queryClient.invalidateQueries(['vehicleData', id])}}
+            style={{marginTop: 10}}
           >
             Back
           </Button>
-          {!readOnly ? (
-            !loading ? (
-              <>
-                <Button
-                  variant='contained'
-                  type='submit'
-                  className={clsx(classes.btn, classes.editButton)}
-                  startIcon={!readOnly ? <NavigateNextRounded /> : <EditIcon />}
-                  disabled={loading}
-                  onClick={loading ? () => null : handleSubmit}
-                >
-                  Save
-                </Button>
-              </>
-            ) : (
-              <div
-                style={{
-                  display: 'flex',
-                  justifyContent: 'flex-end',
-                  width: '90%',
-                  margin: '0 auto',
-                }}
-              >
-                <CircularProgress size={30} />
-              </div>
-            )
-          ) : (
-            <>
-              <div>
-                <Button
-                  variant='contained'
-                  type='submit'
-                  className={clsx(classes.btn, classes.editButton)}
-                  startIcon={!readOnly ? <NavigateNextRounded /> : <EditIcon />}
-                  disabled={loading}
-                  onClick={loading ? () => null : handleEdit}
-                >
-                  Edit
-                </Button>
-              </div>
-            </>
-          )}
         </div>
       </div>
     </div>